test: split frozen Git snapshot coverage by module
